Refactor to introduce service object for entity table check logic
