+++ lib/dfe/analytics.rb
@@ -5,7 +5,8 @@ require 'i18n'
 require 'dfe/analytics/event_schema'
 require 'dfe/analytics/fields'
 require 'dfe/analytics/entities'
-require 'dfe/analytics/entity_processor'
+require 'dfe/analytics/services/entity_table_checks'
+require 'dfe/analytics/shared/service_pattern'
 require 'dfe/analytics/event'
 require 'dfe/analytics/event_matcher'
 require 'dfe/analytics/analytics_job'
